Web_app/requirements.txt
tensorflow==2.5.0
streamlit==1.1.0
pandas==1.3.4
biopython==1.78
Pillow==8.4.0